refactor!: use float instead of double for OpenGL

+++ src/Camera.cpp
@@ -1,35 +1,35 @@
 #include "Camera.h"
 
-PinHoleCamera::PinHoleCamera(Point3d lookFrom, Point3d lookAt, Vector3d up,
-                             double xFov, Vector2i resolution) {
+PinHoleCamera::PinHoleCamera(Point3f lookFrom, Point3f lookAt, Vector3f up,
+                             float xFov, Vector2i resolution) {
    setActualValue(lookFrom, lookAt, up, xFov, resolution);
 }
 
 PinHoleCamera::PinHoleCamera(const Json& cameraJson) {
    const Json& transform_json = cameraJson.at("transform");
-   Vector3d lookAt, lookFrom, up;
+   Vector3f lookAt, lookFrom, up;
    if (transform_json.contains("position")) {
       auto position = transform_json.at("position");
-      lookFrom = Vector3d(position[0], position[1], position[2]);
+      lookFrom = Vector3f(position[0], position[1], position[2]);
    } else {
-      lookFrom = Vector3d(0, 0, 0);
+      lookFrom = Vector3f(0, 0, 0);
    }
 
    if (transform_json.contains("look_at")) {
       auto look_at = transform_json.at("look_at");
-      lookAt = Vector3d(look_at[0], look_at[1], look_at[2]);
+      lookAt = Vector3f(look_at[0], look_at[1], look_at[2]);
    } else {
-      lookAt = Vector3d(0, 1, 0);
+      lookAt = Vector3f(0, 1, 0);
    }
 
    if (transform_json.contains("up")) {
       auto _up = transform_json.at("up");
-      up = Vector3d(_up[0], _up[1], _up[2]);
+      up = Vector3f(_up[0], _up[1], _up[2]);
    } else {
-      up = Vector3d(0, 0, 1);
+      up = Vector3f(0, 0, 1);
    }
 
-   double xFov = 45.f;
+   float xFov = 45.f;
    if (cameraJson.contains("fov")) {
       xFov = cameraJson.at("fov");
    }
@@ -63,15 +63,15 @@ PinHoleCamera::PinHoleCamera(const Json& cameraJson) {
    setActualValue(lookFrom, lookAt, up, xFov, resolution);
 }
 
-void PinHoleCamera::setActualValue(Point3d lookFrom, Point3d lookAt,
-                                   Vector3d up, double xFov,
+void PinHoleCamera::setActualValue(Point3f lookFrom, Point3f lookAt,
+                                   Vector3f up, float xFov,
                                    Vector2i resolution) {
-   double aspectRatio = double(resolution[0]) / resolution[1];
-   double distToFilm = 1.0f / tan(xFov * Transform::PI_F / 360);
+   float aspectRatio = float(resolution[0]) / resolution[1];
+   float distToFilm = 1.0f / tan(xFov * Transform::PI_F / 360);
    view = Transform::getLookAt(lookFrom, lookAt - lookFrom, up);
 
-   double nearPlane = -distToFilm;
-   double farPlane = -100.f;
+   float nearPlane = -distToFilm;
+   float farPlane = -100.f;
    projection = Transform::getPerspective(Transform::AngleDegreeValue(xFov),
                                           aspectRatio, nearPlane, farPlane);
    viewPort = Transform::getViewPort(resolution);
