+++ src/include/Window.h
@@ -8,7 +8,7 @@ class IWindow {
   public:
    virtual void* getNativeWindow() = 0;
    virtual void setNativeWindow(void* window) = 0;
-   virtual void onScroll(double delta) = 0;
+   virtual void onScroll(float delta) = 0;
    virtual void onKey(int key, int scancode, int action, int mods) = 0;
    virtual void onResize(int width, int height) = 0;
    virtual void onClose() = 0;
